refactor(setup): split 746-line setup.ts into SRP modules
Extracted four cohesive modules from the monolithic setup command and
reduced the original file to a 46-line façade preserving the
SetupCommands public API for register.ts/commands.ts callers:

- setup/oauth-pages.ts          — pure HTML for browser OAuth callback
- setup/auth-flow.ts            — auth(), login(), logout(), autoSync
- setup/install-status-line.ts  — Claude Code statusline wiring
- setup/mcp.ts                  — Context7/Linear/Jira MCP registration
- setup/wizard.ts               — start()/setup() orchestration + ASCII art

Zero behavior change — all 1029 tests pass, typecheck and biome clean.
